GRIFFIN: moved compiler specific code to dedicated package
